+++ examples/token-erc20/ponder.config.ts
@@ -1,8 +1,13 @@
 import type { Config } from "@ponder/core";
+import { http } from "viem";
 
 export const config: Config = {
   networks: [
-    { name: "mainnet", chainId: 1, rpcUrl: process.env.PONDER_RPC_URL_1 },
+    {
+      name: "mainnet",
+      chainId: 1,
+      transport: http(process.env.PONDER_RPC_URL_1),
+    },
   ],
   contracts: [
     {