+++ packages/core/src/config/factories.ts
@@ -1,8 +1,8 @@
 import type { Abi, AbiEvent, Address } from "abitype";
 import { getEventSelector, Hex, RpcLog } from "viem";
 
-import type { ResolvedConfig } from "@/config/config";
 import type { Options } from "@/config/options";
+import type { ResolvedConfig } from "@/config/types";
 import { toLowerCase } from "@/utils/lowercase";
 import { getBytesConsumedByParam } from "@/utils/offset";