+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export { PonderApp } from "@/build/handlers";
-export type { Config, ResolvedConfig } from "@/config/config";
+export type { Config, ResolvedConfig } from "@/config/types";
 export type { Block } from "@/types/block";
 export type { ReadOnlyContract } from "@/types/contract";
 export type { Log } from "@/types/log";