+++ packages/create-ponder/src/index.ts
@@ -18,13 +18,13 @@ import { fromSubgraphRepo } from "@/templates/subgraphRepo";
 // @ts-ignore
 import rootPackageJson from "../package.json";
 
-export type Network = {
+export type SerializableNetwork = {
   name: string;
   chainId: number;
-  rpcUrl: string;
+  transport: string;
 };
 
-export type Contract = {
+export type SerializableContract = {
   name: string;
   network: string;
   abi: string;
@@ -32,12 +32,10 @@ export type Contract = {
   startBlock?: number;
 };
 
-export type PartialConfig = {
-  database?: {
-    kind: string;
-  };
-  networks: Network[];
-  contracts: Contract[];
+export type SerializableConfig = {
+  database?: { kind: string };
+  networks: SerializableNetwork[];
+  contracts: SerializableContract[];
 };
 
 export const run = async (
@@ -51,7 +49,7 @@ export const run = async (
   mkdirSync(path.join(rootDir, "abis"), { recursive: true });
   mkdirSync(path.join(rootDir, "src"), { recursive: true });
 
-  let config: PartialConfig;
+  let config: SerializableConfig;
 
   console.log(
     `\nCreating a new Ponder app in ${pico.bold(pico.green(rootDir))}.`
@@ -134,12 +132,15 @@ export const run = async (
   // Write the ponder.config.ts file.
   const finalConfig = `
     import type { Config } from "@ponder/core";
+    import { http } from "viem";
 
     export const config: Config = {
-      networks: ${JSON.stringify(config.networks).replaceAll(
-        /"process.env.PONDER_RPC_URL_(.*?)"/g,
-        "process.env.PONDER_RPC_URL_$1"
-      )},
+      networks: ${JSON.stringify(config.networks)
+        .replaceAll(
+          /"process.env.PONDER_RPC_URL_(.*?)"/g,
+          "process.env.PONDER_RPC_URL_$1"
+        )
+        .replaceAll(/"http\((.*?)\)"/g, "http($1)")},
       contracts: ${JSON.stringify(config.contracts)},
     };
   `;
@@ -165,19 +166,19 @@ export const run = async (
       "scripts": {
         "dev": "ponder dev",
         "start": "ponder start",
+        "codegen": "ponder codegen",
         ${options.eslint ? `"lint": "eslint .",` : ""}
-        "codegen": "ponder codegen"
       },
       "dependencies": {
-        "@ponder/core": "${ponderVersion}",
+        "@ponder/core": "^${ponderVersion}",
       },
       "devDependencies": {
-        ${options.eslint ? `"eslint-config-ponder": "${ponderVersion}",` : ""}
-        ${options.eslint ? `"eslint": "^8.43.0",` : ""}
         "@types/node": "^18.11.18",
         "abitype": "^0.8.11",
+        ${options.eslint ? `"eslint": "^8.43.0",` : ""}
+        ${options.eslint ? `"eslint-config-ponder": "^${ponderVersion}",` : ""}
         "typescript": "^5.1.3",
-        "viem": "^1.2.6"
+        "viem": "^1.2.6",
       }
     }
   `;