+++ packages/create-ponder/src/templates/etherscan.ts
@@ -2,7 +2,7 @@
 import { writeFileSync } from "node:fs";
 import path from "node:path";
 import prettier from "prettier";
-import type { PartialConfig } from "src/index";
+import type { SerializableConfig } from "src/index";
 
 import { getNetworkByEtherscanHostname } from "@/helpers/getEtherscanChainId";
 import { wait } from "@/helpers/wait";
@@ -139,12 +139,12 @@ export const fromEtherscan = async ({
   );
 
   // Build and return the partial ponder config.
-  const config: PartialConfig = {
+  const config: SerializableConfig = {
     networks: [
       {
         name: name,
         chainId: chainId,
-        rpcUrl: `process.env.PONDER_RPC_URL_${chainId}`,
+        transport: `http(process.env.PONDER_RPC_URL_${chainId})`,
       },
     ],
     contracts: [